refactor: domain/user package 재배치
