+++ server/api/src/main/java/vook/server/api/domain/user/logic/UserLogic.java
@@ -6,10 +6,15 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.RequiredArgsConstructor;
 import vook.server.api.domain.user.exception.UserNotFoundException;
-import vook.server.api.domain.user.logic.dto.UserOnboardingCommand;
-import vook.server.api.domain.user.logic.dto.UserRegisterCommand;
-import vook.server.api.domain.user.logic.dto.UserSignUpFromSocialCommand;
-import vook.server.api.domain.user.model.*;
+import vook.server.api.domain.user.model.social_user.SocialUser;
+import vook.server.api.domain.user.model.social_user.SocialUserFactory;
+import vook.server.api.domain.user.model.social_user.SocialUserRepository;
+import vook.server.api.domain.user.model.user.User;
+import vook.server.api.domain.user.model.user.UserFactory;
+import vook.server.api.domain.user.model.user.UserRepository;
+import vook.server.api.domain.user.model.user_info.UserInfo;
+import vook.server.api.domain.user.model.user_info.UserInfoFactory;
+import vook.server.api.domain.user.model.user_info.UserInfoRepository;
 import vook.server.api.globalcommon.annotation.DomainLogic;
 
 import java.util.Optional;
