+++ server/api/src/main/java/vook/server/api/domain/user/logic/UserSignUpFromSocialCommand.java
@@ -1,11 +1,11 @@
-package vook.server.api.domain.user.logic.dto;
+package vook.server.api.domain.user.logic;
 
 import jakarta.validation.constraints.Email;
 import lombok.Builder;
-import vook.server.api.domain.user.model.SocialUser;
-import vook.server.api.domain.user.model.SocialUserFactory;
-import vook.server.api.domain.user.model.User;
-import vook.server.api.domain.user.model.UserFactory;
+import vook.server.api.domain.user.model.social_user.SocialUser;
+import vook.server.api.domain.user.model.social_user.SocialUserFactory;
+import vook.server.api.domain.user.model.user.User;
+import vook.server.api.domain.user.model.user.UserFactory;
 
 @Builder
 public record UserSignUpFromSocialCommand(
