+++ server/api/src/main/java/vook/server/api/domain/user/model/social_user/SocialUserFactory.java
@@ -1,10 +1,11 @@
-package vook.server.api.domain.user.model;
+package vook.server.api.domain.user.model.social_user;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
+import vook.server.api.domain.user.model.user.User;
 
 public interface SocialUserFactory {
-    
+
     SocialUser createForNewOf(
             @NotEmpty String provider,
             @NotEmpty String providerUserId,
