+++ server/api/src/main/java/vook/server/api/domain/user/model/user/UserStatus.java
@@ -1,4 +1,4 @@
-package vook.server.api.domain.user.model;
+package vook.server.api.domain.user.model.user;
 
 public enum UserStatus {
     SOCIAL_LOGIN_COMPLETED,  // 소셜로그인 완료됨
