+++ server/api/src/main/java/vook/server/api/domain/user/model/user_info/Funnel.java
@@ -1,4 +1,4 @@
-package vook.server.api.domain.user.model;
+package vook.server.api.domain.user.model.user_info;
 
 public enum Funnel {
     //X
