+++ server/api/src/main/java/vook/server/api/web/user/usecase/OnboardingUserUseCase.java
@@ -5,9 +5,9 @@ import vook.server.api.domain.template_vocabulary.logic.TemplateVocabularyLogic;
 import vook.server.api.domain.template_vocabulary.model.TemplateTerm;
 import vook.server.api.domain.template_vocabulary.model.TemplateVocabularyType;
 import vook.server.api.domain.user.logic.UserLogic;
-import vook.server.api.domain.user.logic.dto.UserOnboardingCommand;
-import vook.server.api.domain.user.model.Funnel;
-import vook.server.api.domain.user.model.Job;
+import vook.server.api.domain.user.logic.UserOnboardingCommand;
+import vook.server.api.domain.user.model.user_info.Funnel;
+import vook.server.api.domain.user.model.user_info.Job;
 import vook.server.api.domain.vocabulary.logic.term.TermCreateAllCommand;
 import vook.server.api.domain.vocabulary.logic.term.TermLogic;
 import vook.server.api.domain.vocabulary.logic.vocabulary.VocabularyCreateCommand;
