+++ server/api/src/test/java/vook/server/api/domain/user/logic/UserLogicTest.java
@@ -7,13 +7,10 @@ import org.junit.jupiter.api.TestFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
 import vook.server.api.domain.user.exception.*;
-import vook.server.api.domain.user.logic.dto.UserOnboardingCommand;
-import vook.server.api.domain.user.logic.dto.UserRegisterCommand;
-import vook.server.api.domain.user.logic.dto.UserSignUpFromSocialCommand;
-import vook.server.api.domain.user.model.Funnel;
-import vook.server.api.domain.user.model.Job;
-import vook.server.api.domain.user.model.User;
-import vook.server.api.domain.user.model.UserStatus;
+import vook.server.api.domain.user.model.user.User;
+import vook.server.api.domain.user.model.user.UserStatus;
+import vook.server.api.domain.user.model.user_info.Funnel;
+import vook.server.api.domain.user.model.user_info.Job;
 import vook.server.api.globalcommon.exception.ParameterValidateException;
 import vook.server.api.testhelper.IntegrationTestBase;
 import vook.server.api.testhelper.creator.TestUserCreator;
