+++ server/api/src/test/java/vook/server/api/domain/vocabulary/logic/term/TermLogicTest.java
@@ -3,7 +3,7 @@ package vook.server.api.domain.vocabulary.logic.term;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.domain.user.model.User;
+import vook.server.api.domain.user.model.user.User;
 import vook.server.api.domain.vocabulary.exception.TermNotFoundException;
 import vook.server.api.domain.vocabulary.model.term.Term;
 import vook.server.api.domain.vocabulary.model.term.TermRepository;
