+++ server/api/src/test/java/vook/server/api/testhelper/creator/TestUserCreator.java
@@ -4,13 +4,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import vook.server.api.domain.user.logic.UserLogic;
-import vook.server.api.domain.user.logic.dto.UserOnboardingCommand;
-import vook.server.api.domain.user.logic.dto.UserRegisterCommand;
-import vook.server.api.domain.user.logic.dto.UserSignUpFromSocialCommand;
-import vook.server.api.domain.user.model.Funnel;
-import vook.server.api.domain.user.model.Job;
-import vook.server.api.domain.user.model.SocialUser;
-import vook.server.api.domain.user.model.User;
+import vook.server.api.domain.user.logic.UserOnboardingCommand;
+import vook.server.api.domain.user.logic.UserRegisterCommand;
+import vook.server.api.domain.user.logic.UserSignUpFromSocialCommand;
+import vook.server.api.domain.user.model.social_user.SocialUser;
+import vook.server.api.domain.user.model.user.User;
+import vook.server.api.domain.user.model.user_info.Funnel;
+import vook.server.api.domain.user.model.user_info.Job;
 import vook.server.api.web.common.auth.app.TokenService;
 import vook.server.api.web.common.auth.data.GeneratedToken;
