+++ server/api/src/test/java/vook/server/api/web/user/usecase/OnboardingUserUseCaseTest.java
@@ -5,9 +5,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
 import vook.server.api.domain.template_vocabulary.model.TemplateVocabularyType;
-import vook.server.api.domain.user.model.Funnel;
-import vook.server.api.domain.user.model.Job;
-import vook.server.api.domain.user.model.UserRepository;
+import vook.server.api.domain.user.model.user.UserRepository;
+import vook.server.api.domain.user.model.user_info.Funnel;
+import vook.server.api.domain.user.model.user_info.Job;
 import vook.server.api.domain.vocabulary.model.term.TermRepository;
 import vook.server.api.domain.vocabulary.model.vocabulary.UserUid;
 import vook.server.api.domain.vocabulary.model.vocabulary.VocabularyRepository;
