+++ server/api/src/test/java/vook/server/api/web/user/usecase/WithdrawUserUseCaseTest.java
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.domain.user.model.UserRepository;
-import vook.server.api.domain.user.model.UserStatus;
+import vook.server.api.domain.user.model.user.UserRepository;
+import vook.server.api.domain.user.model.user.UserStatus;
 import vook.server.api.domain.vocabulary.model.term.TermRepository;
 import vook.server.api.domain.vocabulary.model.vocabulary.UserUid;
 import vook.server.api.domain.vocabulary.model.vocabulary.Vocabulary;
